Isolate even more Swapchain code.
